+++ lang/syn/src/lib.rs
@@ -592,7 +592,6 @@ pub struct ConstraintGroup {
     pub mutable: Option<ConstraintMut>,
     pub signer: Option<ConstraintSigner>,
     pub owner: Option<ConstraintOwner>,
-    pub rent_exempt: Option<ConstraintRentExempt>,
     pub seeds: Option<ConstraintSeedsGroup>,
     pub executable: Option<ConstraintExecutable>,
     pub has_one: Vec<ConstraintHasOne>,
@@ -636,7 +635,6 @@ pub enum Constraint {
     HasOne(ConstraintHasOne),
     Raw(ConstraintRaw),
     Owner(ConstraintOwner),
-    RentExempt(ConstraintRentExempt),
     Seeds(ConstraintSeedsGroup),
     AssociatedToken(ConstraintAssociatedToken),
     Executable(ConstraintExecutable),
@@ -658,7 +656,6 @@ pub enum ConstraintToken {
     HasOne(Context<ConstraintHasOne>),
     Raw(Context<ConstraintRaw>),
     Owner(Context<ConstraintOwner>),
-    RentExempt(Context<ConstraintRentExempt>),
     Seeds(Context<ConstraintSeeds>),
     Executable(Context<ConstraintExecutable>),
     Close(Context<ConstraintClose>),
@@ -768,12 +765,6 @@ pub struct ConstraintOwner {
 pub struct ConstraintAddress {
     pub address: Expr,
     pub error: Option<Expr>,
-}
-
-#[derive(Debug, Clone)]
-pub enum ConstraintRentExempt {
-    Enforce,
-    Skip,
 }
 
 #[derive(Debug, Clone)]