+++ tests/misc/programs/misc-optional/src/lib.rs
@@ -28,14 +28,6 @@ pub mod misc_optional {
         Ok(())
     }
 
-    pub fn initialize_no_rent_exempt(_ctx: Context<InitializeNoRentExempt>) -> Result<()> {
-        Ok(())
-    }
-
-    pub fn initialize_skip_rent_exempt(_ctx: Context<InitializeSkipRentExempt>) -> Result<()> {
-        Ok(())
-    }
-
     pub fn test_owner(_ctx: Context<TestOwner>) -> Result<()> {
         Ok(())
     }
@@ -284,14 +276,6 @@ pub mod misc_optional {
         Ok(())
     }
 
-    pub fn test_no_rent_exempt(_ctx: Context<NoRentExempt>) -> Result<()> {
-        Ok(())
-    }
-
-    pub fn test_enforce_rent_exempt(_ctx: Context<EnforceRentExempt>) -> Result<()> {
-        Ok(())
-    }
-
     pub fn init_decrease_lamports(ctx: Context<InitDecreaseLamports>) -> Result<()> {
         **ctx
             .accounts
@@ -305,12 +289,6 @@ pub mod misc_optional {
             .as_mut()
             .unwrap()
             .try_borrow_mut_lamports()? += 1;
-        Ok(())
-    }
-
-    pub fn init_if_needed_checks_rent_exemption(
-        _ctx: Context<InitIfNeededChecksRentExemption>,
-    ) -> Result<()> {
         Ok(())
     }